Check permission on sheet usage

+++ pkg/api/handler/sheet_handlers.go
@@ -33,7 +33,13 @@ func (sheetHandler *SheetHandler) CreateSheetHandler(context echo.Context) error
 
 	userId := context.Get("user").(uuid.UUID)
 	sheetRepository := repository.NewSheetRepository(sheetHandler.storage)
-	savedSheet, err := sheetRepository.Create(sheetDto, userId)
+	savedSheet, err := sheetRepository.Create(sheetDto)
+	if err != nil {
+		return api_error.CustomError(context, http.StatusInternalServerError, api_error.SAVING_ERROR, "sheet", sheetDto)
+	}
+
+	sheetUserRepository := repository.NewSheetUserRepository(sheetHandler.storage)
+	_, err = sheetUserRepository.CreateSheetUserRelation(savedSheet.Id, userId, entities.READ, true)
 	if err != nil {
 		return api_error.CustomError(context, http.StatusInternalServerError, api_error.SAVING_ERROR, "sheet", sheetDto)
 	}
@@ -96,9 +102,13 @@ func (sheetHandler *SheetHandler) UpdateSheetHandler(context echo.Context) error
 
 	userId := context.Get("user").(uuid.UUID)
 	sheetRepository := repository.NewSheetRepository(sheetHandler.storage)
-	_, err = sheetRepository.FindByIdAndUserId(id, userId)
+	sheet, err := sheetRepository.FindByIdAndUserId(id, userId)
 	if err != nil {
 		return api_error.CustomError(context, http.StatusBadRequest, api_error.NOT_FOUND, "id", id)
+	}
+
+	if !sheet.Owner && sheet.Permission != entities.EDIT {
+		return api_error.CustomError(context, http.StatusForbidden, api_error.PERMISSION_ERROR, userId)
 	}
 
 	updatedSheet, err := sheetRepository.Update(sheetDto, id)
@@ -118,9 +128,13 @@ func (sheetHandler *SheetHandler) DeleteSheetHandler(context echo.Context) error
 
 	userId := context.Get("user").(uuid.UUID)
 	sheetRepository := repository.NewSheetRepository(sheetHandler.storage)
-	_, err = sheetRepository.FindByIdAndUserId(id, userId)
+	sheet, err := sheetRepository.FindByIdAndUserId(id, userId)
 	if err != nil {
 		return api_error.CustomError(context, http.StatusBadRequest, api_error.NOT_FOUND, "id", id)
+	}
+
+	if !sheet.Owner {
+		return api_error.CustomError(context, http.StatusForbidden, api_error.PERMISSION_ERROR, userId)
 	}
 
 	err = sheetRepository.Delete(id)
